+++ test/fixtures/apps/sequelize/app/model/article.js
@@ -1,6 +1,4 @@
 'use strict';
-const articleJSON = require('./article.json');
-
 module.exports = app => {
   const { STRING, INTEGER } = app.Sequelize;
   const Article = app.model.define('article', {
@@ -22,8 +20,6 @@ module.exports = app => {
     const instance = await Article.findOne({ where: { userId, id } });
     return !!instance;
   };
-
-  Article.settings = articleJSON.settings;
 
   /**
    * 排除未找到错误
@@ -77,8 +73,6 @@ module.exports = app => {
     const [ affectedCount ] = await Article.update(data, { where });
     return { affected: affectedCount };
   };
-
-  Article.remotes = articleJSON.remotes;
 
   return Article;
 };