+++ test/fixtures/apps/sequelize/config/plugin.js
@@ -9,8 +9,4 @@ module.exports = {
     enable: true,
     package: 'egg-connector-rest',
   },
-  validate: {
-    enable: true,
-    package: 'egg-validate',
-  },
 };